+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Support markup in signatures
+- Fix bug where annotations may have been uploaded before comments, causing a failure
 
 ## v0.17.0
 - Always retry on connection issues